# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProjectOps API",
    "description": "Autonomous multi-agent project execution and risk management system. Deterministic scheduling, immutable versioned plans, and append-only audits.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "description": "Get project details including members.",
        "operationId": "get_project_api_v1_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/history/progress": {
      "post": {
        "tags": [
          "history"
        ],
        "summary": "Record Progress Event",
        "description": "Append a task progress event to the audit trail (Rule 5: append-only).",
        "operationId": "record_progress_event_api_v1_projects__project_id__history_progress_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressEventCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "history"
        ],
        "summary": "List Progress History",
        "description": "Retrieve chronological progress history for a project or specific task.",
        "operationId": "list_progress_history_api_v1_projects__project_id__history_progress_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "task_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by task ID",
              "title": "Task Id"
            },
            "description": "Filter by task ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgressEventResponse"
                  },
                  "title": "Response List Progress History Api V1 Projects  Project Id  History Progress Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/decisions": {
      "post": {
        "tags": [
          "history"
        ],
        "summary": "Record Decision",
        "description": "Record a replan or governance decision (Rule 4: Replanning requires approval).",
        "operationId": "record_decision_api_v1_projects__project_id__decisions_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "history"
        ],
        "summary": "List Decisions",
        "description": "Retrieve all governance and replan decisions for a project.",
        "operationId": "list_decisions_api_v1_projects__project_id__decisions_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionResponse"
                  },
                  "title": "Response List Decisions Api V1 Projects  Project Id  Decisions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/risks": {
      "post": {
        "tags": [
          "history"
        ],
        "summary": "Record Risk",
        "description": "Record a detected risk anomaly in the immutable risk register.",
        "operationId": "record_risk_api_v1_projects__project_id__risks_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskEventCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "history"
        ],
        "summary": "List Risks",
        "description": "Retrieve all risk anomaly events recorded for a project.",
        "operationId": "list_risks_api_v1_projects__project_id__risks_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "min_severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by minimum severity",
              "title": "Min Severity"
            },
            "description": "Filter by minimum severity"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RiskEventResponse"
                  },
                  "title": "Response List Risks Api V1 Projects  Project Id  Risks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DecisionCreateRequest": {
        "properties": {
          "decision_type": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Decision Type"
          },
          "summary": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Summary"
          },
          "rationale": {
            "type": "string",
            "maxLength": 4000,
            "title": "Rationale",
            "default": ""
          },
          "decided_by": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Decided By"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          }
        },
        "type": "object",
        "required": [
          "decision_type",
          "summary",
          "decided_by"
        ],
        "title": "DecisionCreateRequest",
        "description": "Payload to record a governance or replan decision."
      },
      "DecisionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "decision_type": {
            "type": "string",
            "title": "Decision Type"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "decided_by": {
            "type": "string",
            "title": "Decided By"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "plan_id",
          "decision_type",
          "summary",
          "rationale",
          "decided_by",
          "recorded_at"
        ],
        "title": "DecisionResponse",
        "description": "Response schema for a decision record."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "daily_capacity_hours": {
            "type": "number",
            "title": "Daily Capacity Hours"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "role",
          "daily_capacity_hours",
          "created_at"
        ],
        "title": "MemberResponse",
        "description": "Response schema for a project member."
      },
...
      "ProgressEventCreateRequest": {
        "properties": {
          "task_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Task Id"
          },
          "new_status": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "New Status"
          },
          "previous_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Status"
          },
          "evidence_notes": {
            "type": "string",
            "maxLength": 2000,
            "title": "Evidence Notes",
            "default": ""
          },
          "recorded_by": {
            "type": "string",
            "maxLength": 255,
            "title": "Recorded By",
            "default": "system"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "new_status"
        ],
        "title": "ProgressEventCreateRequest",
        "description": "Payload to record a task progress event."
      },
      "ProgressEventResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "previous_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Status"
          },
          "new_status": {
            "type": "string",
            "title": "New Status"
          },
          "evidence_notes": {
            "type": "string",
            "title": "Evidence Notes"
          },
          "recorded_by": {
            "type": "string",
            "title": "Recorded By"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "task_id",
          "previous_status",
          "new_status",
          "evidence_notes",
          "recorded_by",
          "recorded_at"
        ],
        "title": "ProgressEventResponse",
        "description": "Response schema for a progress event record."
      },
...
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/MemberResponse"
            },
            "type": "array",
            "title": "Members",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectResponse",
        "description": "Response schema for a project."
      },
      "RiskEventCreateRequest": {
        "properties": {
          "risk_type": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Risk Type"
          },
          "severity": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Severity"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Description"
          },
          "payload_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Json"
          }
        },
        "type": "object",
        "required": [
          "risk_type",
          "severity",
          "description"
        ],
        "title": "RiskEventCreateRequest",
        "description": "Payload to record a detected risk anomaly."
      },
      "RiskEventResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "risk_type": {
            "type": "string",
            "title": "Risk Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "payload_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Json"
          },
          "detected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Detected At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "risk_type",
          "severity",
          "description",
          "payload_json",
          "detected_at"
        ],
        "title": "RiskEventResponse",
        "description": "Response schema for a detected risk anomaly record."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}